refactor: implement specialized category components for different tab types
- Restored `Accordion.tsx` to its original state to keep its responsibility simple.
- Created `ExRStandardCategoryItem` for standard category rendering (used in General Tab).
- Created `ExRRoleCategoryItem` for role-specific category rendering with spawn controls in the header.
- Updated `ExRCategoryList` to switch between these specialized components based on the selected tab ID.
- This approach ensures better separation of concerns by using different components for different rendering requirements instead of a single overloaded component.
